+++ Plugin/src/Content/Unlockables/ShockwaveGalAI.cs
@@ -265,6 +265,7 @@ public class ShockwaveGalAI : NetworkBehaviour, INoiseListener, IHittable
     {
         // Agent.enabled = galState != State.Inactive;
         if (galState == State.Inactive) return;
+        FlySource.volume = Plugin.ModConfig.ConfigShockwaveBotPropellerVolume.Value;
         if (ownerPlayer != null && ownerPlayer.isPlayerDead) ownerPlayer = null;
         HeadPatTrigger.enabled = galState != State.AttackMode && galState != State.Inactive && (ownerPlayer != null && GameNetworkManager.Instance.localPlayerController == ownerPlayer);
         HeadPatTrigger.interactable = galState != State.AttackMode && galState != State.Inactive && (ownerPlayer != null && GameNetworkManager.Instance.localPlayerController == ownerPlayer);
@@ -347,6 +348,10 @@ public class ShockwaveGalAI : NetworkBehaviour, INoiseListener, IHittable
             }
             else
             {
+                if (DetermineIfNeedToDisableAgent(ShockwaveCharger.ChargeTransform.position))
+                {
+                    return false;
+                }
                 Agent.SetDestination(ShockwaveCharger.ChargeTransform.position);
             }
             if (Vector3.Distance(this.transform.position, ShockwaveCharger.ChargeTransform.position) <= Agent.stoppingDistance)