+++ Plugin/src/Patches/ShovelPatch.cs
@@ -5,6 +5,7 @@ using GameNetcodeStuff;
 using System;
 using CodeRebirth.src.Util;
 using CodeRebirth.src.Content.Unlockables;
+using UnityEngine;
 
 namespace CodeRebirth.src.Patches;
 static class ShovelPatch {
@@ -46,10 +47,9 @@ static class ShovelPatch {
     {
         if (!CRWeapon.canBreakTrees) return;
 
-        Plugin.ExtendedLogging("Tree Destroyed: " + CRWeapon.weaponTip.position);
-        RoundManager.Instance.DestroyTreeOnLocalClient(CRWeapon.weaponTip.position);
-
-        if (!Plugin.ModConfig.ConfigFarmingEnabled.Value || random.NextFloat(0f, 1f) >= 0.02f) return;
+		int num = Physics.OverlapSphereNonAlloc(CRWeapon.weaponTip.position, 5f, RoundManager.Instance.tempColliderResults, 33554432, QueryTriggerInteraction.Ignore);
+        if (!Plugin.ModConfig.ConfigFarmingEnabled.Value || random.NextFloat(0f, 1f) >= 0.02f || num <= 0) return;
+		RoundManager.Instance.DestroyTreeOnLocalClient(CRWeapon.weaponTip.position);
         Plugin.ExtendedLogging("Tree Destroyed with luck");
         CodeRebirthUtils.Instance.SpawnScrap(UnlockableHandler.Instance.PlantPot.Seed, CRWeapon.weaponTip.position, false, true, 5);
     }